Fix errors when public page is pending and error
